Create CRUD for entities

+++ BE/ecommerce/src/main/java/com/ecommerce/demo/repositories/RoleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ecommerce.demo.data.entities.RoleEntity;
 
-public interface RoleRepository extends JpaRepository<RoleEntity, Integer>{
+public interface RoleRepository extends JpaRepository<RoleEntity, Long>{
 	
 }
